Rename package from `org.rubyevents.RubyEvents` to `org.rubyevents.app`
Android package name parts should not start with an uppercase letter.
I initially wanted to change it to "org.rubyevents.rubyevents", but the redundancy seemed unnecessary.
It's a common practice to use "app" as a suffix so I went for that.
