+++ app/src/androidTest/java/org/rubyevents/app/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package org.rubyevents.RubyEvents
+package org.rubyevents.app
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
   fun useAppContext() {
     // Context of the app under test.
     val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-    assertEquals("org.rubyevents.RubyEvents", appContext.packageName)
+    assertEquals("org.rubyevents.app", appContext.packageName)
   }
 }
