+++ app/src/test/java/org/rubyevents/app/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package org.rubyevents.RubyEvents
+package org.rubyevents.app
 
 import org.junit.Test
